% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appendR.R
\name{appendR}
\alias{appendR}
\title{Appends parameter estimates into list of data.table ...}
\usage{
appendR(result, modelFit, all = FALSE, noIntercept = FALSE)
}
\arguments{
\item{result}{List of data.tables that contain results}

\item{modelFit}{Model fit (works with glm and lm for sure)}

\item{all}{Indicator. If TRUE, include Estimate, SE,
Test Statistic, and p-Value. If FALSE, include
Estimate and p-Value only.}

\item{noIntercept}{If TRUE, exclude Intercept estimates.}
}
\value{
plot
}
\description{
Appends parameter estimates into list of data.table ...
}
