% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtsample.R
\name{dtSample}
\alias{dtSample}
\title{Sample ids from a data.table}
\usage{
dtSample(dt, n, sID, repVal = FALSE)
}
\arguments{
\item{dt}{A data.table.}

\item{n}{An integer.}

\item{sID}{A string.}

\item{repVal}{True or False}
}
\value{
A vector
}
\description{
Sample ids from a data.table
}

