% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.R
\name{estciStr}
\alias{estciStr}
\title{Create a data.table of ci strings}
\usage{
estciStr(object, digits = 3)
}
\arguments{
\item{object}{a data.table created by function confintExp}

\item{digits}{number of digits to be printed}
}
\value{
A data.table that stores a confidence interval in string format
}
\description{
Create a data.table of ci strings
}

