% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdir.R
\name{fdir}
\alias{fdir}
\title{Create a filename  with the current working directory}
\usage{
fdir(directory, filename)
}
\arguments{
\item{directory}{A string.}

\item{filename}{A string.}
}
\value{
A string
}
\description{
Create a filename  with the current working directory
}

