% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myTableOne.R
\name{myTableOne}
\alias{myTableOne}
\title{Creates Table 1 data}
\usage{
myTableOne(dT, digits = 2)
}
\arguments{
\item{dT}{data.table with raw data that needs to be summarized. Exclude
any fields that should not be in Table 1, such as id information}

\item{digits}{Specify precision}
}
\value{
data.table
}
\description{
Creates Table 1 data
}
