% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noAxisTitle.R
\name{noAxisTitle}
\alias{noAxisTitle}
\title{Removes axis titles quickly ...}
\usage{
noAxisTitle(arg = "both")
}
\arguments{
\item{axis}{that is referenced: x, y, or both}
}
\value{
ggplot theme
}
\description{
Removes axis titles quickly ...
}
