% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packCountDL.R
\name{pkgDLcount}
\alias{pkgDLcount}
\title{Counts number of simstudy downloads ...}
\usage{
pkgDLcount(startdate = "2016-06-21", package = "simstudy")
}
\arguments{
\item{startdate}{Character date that represents point
from which to start counting.}

\item{package}{Package name}
}
\value{
integer
}
\description{
Counts number of simstudy downloads ...
}
