% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packPlotDL.R
\name{pkgDLplot}
\alias{pkgDLplot}
\title{PLots number of simstudy downloads ...}
\usage{
pkgDLplot(package = "simstudy", by_period = "day", cumulative = TRUE)
}
\arguments{
\item{package}{String}

\item{by_period}{"day" (default), "month", or "year"}

\item{cumulative}{TRUE (default) or FALSE}
}
\value{
plot
}
\description{
PLots number of simstudy downloads ...
}
