% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myPrintTable.R
\name{prepTable}
\alias{prepTable}
\title{Prepare table for printing}
\usage{
prepTable(tab, missing = FALSE)
}
\arguments{
\item{tab}{data.table created by myTableOne ...}

\item{missing}{Flag to indicate if count of missing data should be included in table}
}
\value{
data.table
}
\description{
Prepare table for printing
}
