% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readexcel.R
\name{readExcel}
\alias{readExcel}
\title{Reads from clipboard into data.frame: windows}
\usage{
readExcel(header = TRUE, ...)
}
\arguments{
\item{header}{True / False.}

\item{...}{extra arguments}
}
\value{
A data frame
}
\description{
Wrapper for read.table that reads from the clipboard into a data.frame with
or without the header - works only on Windows.
}

