% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwExcel.R
\name{readExcelMac}
\alias{readExcelMac}
\title{Reads from clipboard into data.frame: mac}
\usage{
readExcelMac(header = TRUE, ...)
}
\arguments{
\item{header}{True or False.}

\item{...}{extra arguments}
}
\value{
A data frame
}
\description{
Wrapper for read.table that reads from the clipboard into a data.frame with
or without the header - works only on a Mac.
}

