% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replaceInFiles.R
\name{replaceInFiles}
\alias{replaceInFiles}
\title{Find and replace text in all files in a directory}
\usage{
replaceInFiles(directory, oldtext, newtext)
}
\arguments{
\item{directory}{Location of files to be changed}

\item{oldtext}{Current text}

\item{newtext}{New text}
}
\value{
Nothing
}
\description{
Find and replace text in all files in a directory
}
