% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rndTidy.R
\name{rndTidy}
\alias{rndTidy}
\title{Appends parameter estimates into list of data.table ...}
\usage{
rndTidy(modelFit, digits = 2)
}
\arguments{
\item{modelFit}{Model fit (works with glm and lm for sure) object}

\item{digits}{Number of digits to round to}
}
\value{
plot
}
\description{
Appends parameter estimates into list of data.table ...
}
