% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_condition.R
\name{show_condition}
\alias{show_condition}
\title{Screens functions for messages}
\usage{
show_condition(code)
}
\arguments{
\item{code}{Function call}
}
\value{
If the function results in a message, the function returns a string
indicating "error", "warning", or "message"; otherwise the object of the
function call is returned.
}
\description{
Checks to see if function call returns error
}

