% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumdata.R
\name{sumData}
\alias{sumData}
\title{Create summary - n, mean, sd, min, max of all numerical data}
\usage{
sumData(dt, threshold = NULL)
}
\arguments{
\item{dt}{a data.table with at least one numeric column}
}
\value{
A data.table with summary stats: n, min, max, mean, sd
}
\description{
Create summary - n, mean, sd, min, max of all numerical data
}

