% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_theme.R
\name{theme_ksg}
\alias{theme_ksg}
\title{Sets "theme" for ggplot2}
\usage{
theme_ksg(panelback = "white")
}
\arguments{
\item{panelback}{Color of panel background}
}
\value{
ggplot theme
}
\description{
Resets defaults for themes ...
}

