% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitsToInt.R
\name{bitsToInt}
\alias{bitsToInt}
\title{Converts vector of 1's and 0's to integer ...}
\usage{
bitsToInt(x)
}
\arguments{
\item{x}{Vector of 1's or zero's}
}
\value{
Integer value of vector
}
\description{
Converts vector of 1's and 0's to integer ...
}

