% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitsToInt.R
\name{bitsToInt}
\alias{bitsToInt}
\title{Convert bits it integer ...}
\usage{
bitsToInt(x)
}
\arguments{
\item{x}{Vector of 1's or zero's}
}
\value{
Integer value of vector
}
\description{
Convert bits it integer ...
}
