% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confintExp.R
\name{confintExp}
\alias{confintExp}
\title{Estimate confidence intervals from GLM}
\usage{
confintExp(object)
}
\arguments{
\item{object}{A glm fitted model object.}
}
\value{
Data.table that stores a confidence interval
}
\description{
Estimate confidence intervals from GLM
}
