% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packPlotDL.R
\name{pkgDLplot}
\alias{pkgDLplot}
\title{PLots number of simstudy downloads ...}
\usage{
pkgDLplot(period = 7, package = "simstudy", cumulative = FALSE)
}
\arguments{
\item{period}{Integer value length of period, defaults to 7 (week)}
}
\value{
plot
}
\description{
PLots number of simstudy downloads ...
}
